site/Backend/requirements.txt
absl-py==2.3.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.73.0
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==2.10.0
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
ml_dtypes==0.5.1
mtcnn==1.0.0
namex==0.1.0
numpy==1.23.5
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==3.20.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2025.2
requests==2.32.4
retina-face==0.0.17
six==1.17.0
soupsieve==2.7
spotipy==2.25.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorflow==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
